refactor: Refactor component imports and restructure theme provider; migrate ProfileSidebar and TableFilter to new paths
