+++ src/app/dashboard/allmembers/member/page.tsx
@@ -1,4 +1,4 @@
-import { ProfileTabs } from "@/components/ProfileTabs";
+import { ProfileTabs } from "@/components/pages/allmembers/profiles/ProfileTabs";
 
 export default function page() {
   return (
